HSEARCH-3792 Move ElasticsearchJsonSyntaxHelper to a package that is not tied to search
So that we can use it in other packages.
