+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/util/impl/Elasticsearch56JsonSyntaxHelper.java
@@ -4,17 +4,16 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.search.impl;
+package org.hibernate.search.backend.elasticsearch.util.impl;
 
-public class Elasticsearch7JsonSyntaxHelper implements ElasticsearchJsonSyntaxHelper {
-
+public class Elasticsearch56JsonSyntaxHelper extends Elasticsearch7JsonSyntaxHelper {
 	@Override
 	public String getTermAggregationOrderByTermToken() {
-		return "_key";
+		return "_term"; // _key in ES6.0+
 	}
 
 	@Override
 	public boolean useOldSortNestedApi() {
-		return false;
+		return true;
 	}
 }
