+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/util/impl/Elasticsearch7JsonSyntaxHelper.java
@@ -4,16 +4,17 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.search.impl;
+package org.hibernate.search.backend.elasticsearch.util.impl;
 
-public class Elasticsearch56JsonSyntaxHelper extends Elasticsearch7JsonSyntaxHelper {
+public class Elasticsearch7JsonSyntaxHelper implements ElasticsearchJsonSyntaxHelper {
+
 	@Override
 	public String getTermAggregationOrderByTermToken() {
-		return "_term"; // _key in ES6.0+
+		return "_key";
 	}
 
 	@Override
 	public boolean useOldSortNestedApi() {
-		return true;
+		return false;
 	}
 }
